Prototype script to deal with dates in data
The Campbell Scientific data logger outputs the year, day of month, and
time in minutes in separate columns. However, to plot any data in Excel
we really need a single column with a datetime. This script is a first
sketch at adding a column with the date from each of these distinct data
fields.